+++ tools/corda-runtime-gradle-plugin/src/endToEndTest/kotlin/net/corda/gradle/plugin/queries/QueriesSmokeTest.kt
@@ -1,59 +1,23 @@
 package net.corda.gradle.plugin.queries
 
-import kotlinx.coroutines.runBlocking
 import net.corda.gradle.plugin.EndToEndTestBase
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import java.io.File
-import java.util.concurrent.TimeUnit
 
 class QueriesSmokeTest : EndToEndTestBase() {
     companion object {
-        private val composeFile = File(this::class.java.getResource("/config/combined-worker-compose.yml")!!.toURI())
-
         @JvmStatic
         @BeforeAll
         fun startCombinedWorker() {
-            val cordaProcessBuilder = ProcessBuilder(
-                "docker",
-                "compose",
-                "-f",
-                composeFile.absolutePath,
-                "-p",
-                "corda-cluster",
-                "up",
-                "--pull",
-                "missing",
-                "--quiet-pull",
-                "--detach"
-            )
-            cordaProcessBuilder.environment()["CORDA_RUNTIME_VERSION"] = "5.3.0.0-HC01"
-            cordaProcessBuilder.redirectErrorStream(true)
-            val cordaProcess = cordaProcessBuilder.start()
-            cordaProcess.inputStream.transferTo(System.out)
-            cordaProcess.waitFor(10, TimeUnit.SECONDS)
-            if (cordaProcess.exitValue() != 0) throw IllegalStateException("Failed to start Corda cluster using docker compose")
-            runBlocking { waitUntilRestApiIsAvailable() }
+            startCompose()
         }
 
         @JvmStatic
         @AfterAll
         fun stopCombinedWorker() {
-            val cordaProcessBuilder = ProcessBuilder(
-                "docker",
-                "compose",
-                "-f",
-                composeFile.absolutePath,
-                "-p",
-                "corda-cluster",
-                "down",
-            )
-            cordaProcessBuilder.redirectErrorStream(true)
-            val cordaStopProcess = cordaProcessBuilder.start()
-            cordaStopProcess.inputStream.transferTo(System.out)
-            cordaStopProcess.waitFor(30, TimeUnit.SECONDS)
+            stopCompose()
         }
     }
